+++ Tests/SwiftLintFrameworkTests/ConfigurationAliasesTests.swift
@@ -1,5 +1,3 @@
-import Foundation
-import SourceKittenFramework
 @testable import SwiftLintFramework
 import XCTest