+++ Tests/SwiftLintFrameworkTests/GlobTests.swift
@@ -1,5 +1,4 @@
 import Foundation
-import SourceKittenFramework
 @testable import SwiftLintFramework
 import XCTest